fix: fix CR comments
